backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-multipart>=0.0.5
pyPDF2>=2.10.5
transformers>=4.18.0
torch>=1.10.0
sentence-transformers>=2.2.0
numpy>=1.21.0
elasticsearch>=7.14.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0